+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog for NeoFS Contract
 ### Added
 - Structured containers to Container contract storage (#534)
 - `addStructs`, `createV2` and `getInfo` methods to Container contract (#534)
+- NEP-11 support by Container contract (#498)
 
 ### Changed